# docs/openapi/loyalty-admin-bff.yaml
openapi: 3.0.3
info:
  title: loyalty-admin-bff — Bank Employee Portal API
  version: '0.1.0'
  description: |
    Employee-facing edge API for the Rochallor Loyalty Platform, consumed by the Bank Employee
    Portal (BEP) via AWS API Gateway under `/api/loyalty/admin/*`. Aggregates `loyalty-core`,
    `loyalty-earning`, `loyalty-redemption`, and `loyalty-campaign` (Arch §4.2.2).

    **Auth:** the RDB Authentication Service verifies the employee token and injects the actor into the
    request as the `userId` + comma-separated `roles` query parameters before it reaches the BFF; the BFF
    does no token handling. Loyalty roles gate operations (Arch §7.1): `loyalty-cs-maker`,
    `loyalty-cs-checker`, `loyalty-campaign-manager`, `loyalty-fraud-ops`, `loyalty-admin`,
    `loyalty-readonly`. The employee `userId` is distinct from the `customerId` (CIF) that member-scoped
    endpoints act on. Per-Program scope is enforced via `PROGRAM_ADMIN` assignment; a maker-checker action
    in Program A cannot affect Program B.

    **Approvals:** Loyalty does **not** implement maker-checker. Money-equivalent and
    economic-config changes (adjustments, rule/reward/tier/T&Cs activation) are raised via
    `POST /approval-requests` and applied via `POST /approval-requests/{id}/confirm` once the bank's
    **BEP Approval Workflow** approves. BEP owns routing, Job Roles, 4-eyes, and caps; Loyalty stores
    only a `bepApprovalRef`. The negative-balance policy is unchanged.

...
  contact:
...
servers:
...
security:
  - employeeJwt: []
...
paths:
...
  /reward-types:
    get:
      tags: [Rewards]
      operationId: listRewardTypes
      summary: Platform Reward Type catalogue (seeded).
      responses:
        '200':
          description: Reward types.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RewardType' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /programs/{programId}/rewards:
    get:
      tags: [Rewards]
      operationId: listRewardsAdmin
      summary: Reward catalogue for a Program (all statuses).
      parameters: [ { $ref: '#/components/parameters/ProgramId' } ]
      responses:
        '200':
          description: Rewards.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminReward' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Rewards]
      operationId: createReward
      summary: Author a new Reward (DRAFT).
      parameters: [ { $ref: '#/components/parameters/ProgramId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rewardTypeCode, name, pointCost]
              properties:
                rewardTypeCode: { type: string }
                name: { type: string }
                pointCost: { type: integer, format: int64 }
                fulfillmentParams: { type: object, additionalProperties: true }
                eligibility: { type: object, additionalProperties: true }
                inventoryTotal: { type: integer, format: int64, nullable: true }
      responses:
        '201':
          description: Reward created.
          content: { application/json: { schema: { $ref: '#/components/schemas/AdminReward' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /rewards/{rewardId}:
    patch:
      tags: [Rewards]
      operationId: updateRewardStatus
      summary: Activate or archive a Reward.
      description: Activation (→ ACTIVE) and `point_cost` changes are economic and **approval-gated** — they raise an `ApprovalRequest` of type `REWARD_CHANGE` and apply on `confirm`. `ARCHIVED` applies directly.
      parameters: [ { $ref: '#/components/parameters/RewardId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [ACTIVE, ARCHIVED] }
      responses:
        '200':
          description: Updated.
          content: { application/json: { schema: { $ref: '#/components/schemas/AdminReward' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    employeeJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Employee JWT from RDB Authentication Service (Keycloak), business-employee-portal realm. Roles gate operations (Arch §7.1).
  parameters:
...
    ProgramId:
      name: programId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
    RewardId:
      name: rewardId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT.
      content: { application/json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Forbidden:
      description: Authenticated but lacks the required role / Program scope.
      content: { application/json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
  schemas:
    ProgramRef:
      type: object
      required: [programId, programCode]
      properties:
        programId: { type: integer, format: int64 }
        programCode: { type: string, maxLength: 32, example: RDB_REWARDS }
...
    RewardType:
      type: object
      properties:
        rewardTypeCode: { type: string, example: CASHBACK }
        displayName: { type: string }
        fulfillmentAdapterClass: { type: string }
        parameterSchema: { type: object, additionalProperties: true }
    AdminReward:
      allOf:
        - $ref: '#/components/schemas/ProgramRef'
        - type: object
          properties:
            rewardId: { type: integer, format: int64 }
            rewardTypeCode: { type: string }
            rewardRevision: { type: integer }
            name: { type: string }
            pointCost: { type: integer, format: int64 }
            status: { type: string, enum: [DRAFT, ACTIVE, ARCHIVED] }
            inventoryTotal: { type: integer, format: int64, nullable: true }
            inventoryRemaining: { type: integer, format: int64, nullable: true }
...
    Problem:
      type: object
      description: RFC 7807 problem detail.
      properties:
        type: { type: string, format: uri, default: 'about:blank' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string, description: 'Loyalty error code, e.g. CAP_EXCEEDED, CHECKER_EQUALS_MAKER, DSL_INVALID.' }